+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Edit Inventory Instances: Display a Save & keep editing button. Refs UIIN-2457.
 * Edit Inventory Holdings: Display a Save & keep editing button. Refs UIIN-2404.
 * "holdings-storage" API version upgrade. Refs UIIN-2926.
+* Edit Inventory Items: Display a Save & keep editing button. Refs UIIN-2456.
 
 ## [11.0.4](https://github.com/folio-org/ui-inventory/tree/v11.0.4) (2024-04-30)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v11.0.3...v11.0.4)